tests/wpt/web-platform-tests/css/css-animations/parsing/animation-shorthand.tentative.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.csswg.org/css-animations/#propdef-animation">
<link rel="help" href="https://drafts.csswg.org/scroll-animations-1/">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/shorthand-testcommon.js"></script>
<script>
// TODO(https://github.com/w3c/csswg-drafts/issues/8054): When support for
// animation-delay-start and -end is added to the animation shorthand, this
// file should just merge with animation-shorthand.html.
test_shorthand_value('animation', 'anim paused both reverse 4 1s -3s cubic-bezier(0, -2, 1, 3)', {
  'animation-duration': '1s',
  'animation-timing-function': 'cubic-bezier(0, -2, 1, 3)',
  'animation-delay-start': '-3s',
  'animation-delay-end': '0s',
  'animation-iteration-count': '4',
  'animation-direction': 'reverse',
  'animation-fill-mode': 'both',
  'animation-play-state': 'paused',
  'animation-name': 'anim',
  'animation-timeline': 'auto',
  'animation-range-start': 'normal',
  'animation-range-end': 'normal',
});
</script>
